Corrected jquery.ui tab role bug
Easiest way to see the change is look at the standard jquery.ui.1.12.1 and search for "presentation".  There's only one in the code, it's a role attribute added by the tabs functionality.  In the UHI version we have replaced "presentation" with "button".  This is a known bug (https://github.com/Geeklog-Core/geeklog/issues/693).